Towards 64-bit pageids.  bufferManager can handle 64-bit offsets, but recordid.page is still 32 bit...

+++ test/lladd/check_indirect.c
@@ -132,7 +132,7 @@ START_TEST(indirectAlloc) {
   
 
 
-  printf("{page = %d, slot = %d, size = %lld}\n", rid.page, rid.slot, (long long int)rid.size);
+  printf("{page = %lld, slot = %d, size = %lld}\n", (int64_t)rid.page, rid.slot, (int64_t)rid.size);
 
   releasePage(p);
 
@@ -156,7 +156,7 @@ START_TEST(indirectAlloc) {
  
 
 
-  printf("{page = %d, slot = %d, size = %lld}\n", rid.page, rid.slot, (long long int)rid.size);
+  printf("{page = %lld, slot = %d, size = %lld}\n", (int64_t)rid.page, rid.slot, (int64_t)rid.size);
 
   releasePage(p);
